LPS-55414 Split tests into service and local service tests
